+++ src/client/datascience/jupyter/kernels/jupyterKernelPromiseFailedError.ts
